+++ duckscript_sdk/src/sdk/std/process/spawn/mod.rs
@@ -1,0 +1,46 @@
+use crate::utils::{exec, pckg};
+use duckscript::types::command::{Command, CommandResult};
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "Spawn")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["spawn".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
+    }
+
+    fn run(&self, arguments: Vec<String>) -> CommandResult {
+        match exec::spawn(&arguments, false, false, 0) {
+            Ok(child) => {
+                let pid = child.id();
+
+                CommandResult::Continue(Some(pid.to_string()))
+            }
+            Err(error) => CommandResult::Error(error),
+        }
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}