+++ duckscript_sdk/src/sdk/std/process/spawn/mod_test.rs
@@ -1,0 +1,22 @@
+use super::*;
+use crate::test;
+use crate::test::CommandValidation;
+
+#[test]
+fn common_functions() {
+    test::test_common_command_functions(create(""));
+}
+
+#[test]
+fn run_no_args() {
+    test::run_script_and_error(vec![create("")], "out = spawn", "out");
+}
+
+#[test]
+fn run_valid() {
+    test::run_script_and_validate(
+        vec![create("")],
+        "out = spawn echo test",
+        CommandValidation::PositiveNumber("out".to_string()),
+    );
+}