+++ duckscript_sdk/src/utils/exec.rs
@@ -1,5 +1,4 @@
-use std::process::Command;
-use std::process::Stdio;
+use std::process::{Child, Command, Stdio};
 
 #[cfg(test)]
 #[path = "./exec_test.rs"]
@@ -11,6 +10,49 @@ pub(crate) fn exec(
     allow_input: bool,
     start_index: usize,
 ) -> Result<(String, String, i32), String> {
+    let mut command = create_command(arguments, print_output, allow_input, start_index)?;
+
+    match command.output() {
+        Ok(ref output) => {
+            let stdout = String::from_utf8_lossy(&output.stdout).into_owned();
+            let stderr = String::from_utf8_lossy(&output.stderr).into_owned();
+            let exit_code = match output.status.code() {
+                Some(value) => value,
+                None => {
+                    return Err(format!(
+                        "Unable to extract exit code for command: {}",
+                        &arguments[0]
+                    )
+                    .to_string());
+                }
+            };
+
+            Ok((stdout, stderr, exit_code))
+        }
+        Err(error) => Err(error.to_string()),
+    }
+}
+
+pub(crate) fn spawn(
+    arguments: &Vec<String>,
+    print_output: bool,
+    allow_input: bool,
+    start_index: usize,
+) -> Result<Child, String> {
+    let mut command = create_command(arguments, print_output, allow_input, start_index)?;
+
+    match command.spawn() {
+        Ok(child) => Ok(child),
+        Err(error) => Err(error.to_string()),
+    }
+}
+
+fn create_command(
+    arguments: &Vec<String>,
+    print_output: bool,
+    allow_input: bool,
+    start_index: usize,
+) -> Result<Command, String> {
     if arguments.len() <= start_index {
         Err("Command not provided.".to_string())
     } else {
@@ -31,24 +73,6 @@ pub(crate) fn exec(
             command.stdout(Stdio::inherit()).stderr(Stdio::inherit());
         }
 
-        match command.output() {
-            Ok(ref output) => {
-                let stdout = String::from_utf8_lossy(&output.stdout).into_owned();
-                let stderr = String::from_utf8_lossy(&output.stderr).into_owned();
-                let exit_code = match output.status.code() {
-                    Some(value) => value,
-                    None => {
-                        return Err(format!(
-                            "Unable to extract exit code for command: {}",
-                            &arguments[0]
-                        )
-                        .to_string());
-                    }
-                };
-
-                Ok((stdout, stderr, exit_code))
-            }
-            Err(error) => Err(error.to_string()),
-        }
+        Ok(command)
     }
 }