+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.2.1
 
+* New spawn command.
 * Update build makefile.
 * Runtime - Add support for **Any** state type.
 * SDK - Use fsio crate for file system apis.